+++ library/android/src/main/java/io/mehow/ruler/Ruler.kt
@@ -48,10 +48,10 @@ public object Ruler : LengthConverter, LengthFormatter {
     formatterFactories -= factory
   }
 
-  override fun Length<*>.format(context: Context, unitSeparator: String): String? = installedFormatterFactories
+  override fun Length<*>.format(unitSeparator: String, context: Context): String? = installedFormatterFactories
       .asSequence()
       .mapNotNull { factory -> factory.create(this, unitSeparator) }
-      .map { formatter -> with(formatter) { format(context, unitSeparator) } }
+      .map { formatter -> with(formatter) { format(unitSeparator, context) } }
       .firstOrNull()
 
   private val builtInImperialCountryCodes = setOf("US", "LR", "MM")
@@ -88,6 +88,6 @@ public fun Length<*>.format(
   formatter: LengthFormatter = Ruler,
 ): String {
   val length = converter?.run { convert(context) } ?: this
-  val text = with(formatter) { length.format(context, unitSeparator) }
+  val text = with(formatter) { length.format(unitSeparator, context) }
   return checkNotNull(text) { "Failed to format length: $length" }
 }